+++ src/main/java/com/ardaslegends/presentation/api/BattleRestController.java
@@ -2,7 +2,7 @@ package com.ardaslegends.presentation.api;
 
 import com.ardaslegends.domain.war.Battle;
 import com.ardaslegends.presentation.api.response.war.BattleResponse;
-import com.ardaslegends.service.dto.war.CreateBattleDto;
+import com.ardaslegends.service.dto.war.battle.CreateBattleDto;
 import com.ardaslegends.service.war.BattleService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
