+++ src/test/java/com/ardaslegends/service/BattleServiceTest.java
@@ -7,7 +7,7 @@ import com.ardaslegends.domain.war.War;
 import com.ardaslegends.domain.war.WarParticipant;
 import com.ardaslegends.repository.BattleRepository;
 import com.ardaslegends.repository.war.WarRepository;
-import com.ardaslegends.service.dto.war.CreateBattleDto;
+import com.ardaslegends.service.dto.war.battle.CreateBattleDto;
 import com.ardaslegends.service.exceptions.logic.army.ArmyServiceException;
 import com.ardaslegends.service.exceptions.logic.war.BattleServiceException;
 import com.ardaslegends.service.war.BattleService;
